# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Temporary email service with identity-based authentication using BIP39 mnemonics",
        "title": "CoreSend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/health": {
            "get": {
                "description": "Check API and services health status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "operationId": "healthCheck",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthResponse"
                        }
                    },
                    "503": {
                        "description": "Service unavailable - Redis disconnected",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/inbox/{address}": {
            "get": {
                "security": [
                    {
                        "SignatureAuth": []
                    }
                ],
                "description": "Retrieve all emails for a specific address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbox"
                ],
                "summary": "Get inbox emails",
                "operationId": "getInbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address to retrieve emails for",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.InboxResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SignatureAuth": []
                    }
                ],
                "description": "Delete all emails for a specific address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbox"
                ],
                "summary": "Clear entire inbox",
                "operationId": "clearInbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address to clear inbox for",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/inbox/{address}/{emailId}": {
            "get": {
                "security": [
                    {
                        "SignatureAuth": []
                    }
                ],
                "description": "Retrieve a specific email by ID for an address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbox"
                ],
                "summary": "Get single email",
                "operationId": "getEmail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email ID",
                        "name": "emailId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.EmailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SignatureAuth": []
                    }
                ],
                "description": "Delete a specific email by ID for an address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbox"
                ],
                "summary": "Delete single email",
                "operationId": "deleteEmail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email ID",
                        "name": "emailId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/register/{address}": {
            "post": {
                "security": [
                    {
                        "SignatureAuth": []
                    }
                ],
                "description": "Register a derived hex address to receive emails for the next 24 hours",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbox"
                ],
                "summary": "Register address for inbound mail",
                "operationId": "registerAddress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex-encoded address to register",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid address format or missing address",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.DeleteResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 5
                },
                "deleted": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "api.EmailResponse": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "body": {
                    "type": "string",
                    "example": "This is the email body content"
                },
                "from": {
                    "type": "string",
                    "example": "sender@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "received_at": {
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "subject": {
                    "type": "string",
                    "example": "Hello World"
                },
                "to": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2@coresend.dev"
                    ]
                }
            }
        },
        "api.ErrorDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_ADDRESS"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string",
                    "example": "Address is required"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.ErrorDetails"
                }
            }
        },
        "api.HealthResponse": {
            "type": "object",
            "properties": {
                "services": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "connected"
                }
            }
        },
        "api.InboxResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
                },
                "count": {
                    "type": "integer",
                    "example": 5
                },
                "email": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2@coresend.dev"
                },
                "emails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.EmailResponse"
                    }
                }
            }
        },
        "api.RegisterResponse": {
            "type": "object",
            "required": [
                "address",
                "expires_in",
                "registered"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
                },
                "expires_in": {
                    "type": "integer",
                    "example": 86400
                },
                "registered": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}